kb/scraping/agevolazioni/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
google-cloud-storage==2.14.0
google-auth==2.25.2
lxml==4.9.3
python-dotenv==1.0.0
